--- src/main/java/uk/ac/ebi/embl/gff3tools/sequence/SubmissionType.java
@@ -1,6 +1,0 @@
-package uk.ac.ebi.embl.gff3tools.sequence;
-
-public enum SubmissionType{
-    FASTA, 
-    PLAIN_SEQUENCE,
-}
