+++ src/main/java/uk/ac/ebi/embl/gff3tools/sequence/fasta/headerutils/FastaHeader.java
@@ -8,7 +8,7 @@
  * CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package uk.ac.ebi.embl.gff3tools.fasta.headerutils;
+package uk.ac.ebi.embl.gff3tools.sequence.fasta.headerutils;
 
 import com.fasterxml.jackson.annotation.JsonAlias;
 import com.fasterxml.jackson.annotation.JsonProperty;
