+++ src/main/java/uk/ac/ebi/embl/gff3tools/sequence/fasta/headerutils/ParsedHeader.java
@@ -8,9 +8,12 @@
  * CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package uk.ac.ebi.embl.gff3tools.fasta;
+package uk.ac.ebi.embl.gff3tools.sequence.fasta.headerutils;
 
-public enum IdType {
-    ACCESSIONID,
-    SUBMITTERID
+import lombok.Value;
+
+@Value
+public class ParsedHeader {
+    String id;
+    FastaHeader header;
 }
