+++ src/main/java/uk/ac/ebi/embl/gff3tools/sequence/wrappers/SubmissionType.java
@@ -1,0 +1,6 @@
+package uk.ac.ebi.embl.gff3tools.sequence.wrappers;
+
+enum SubmissionType{
+    FASTA, 
+    PLAIN_SEQUENCE,
+}
