+++ src/test/java/uk/ac/ebi/embl/gff3tools/sequence/fasta/JsonHeaderFastaReaderTest.java
@@ -8,7 +8,7 @@
  * CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package uk.ac.ebi.embl.gff3tools.fasta;
+package uk.ac.ebi.embl.gff3tools.sequence.fasta;
 
 import static org.junit.jupiter.api.Assertions.*;
